Local Storage - list Players

+++ src/storage/group/groupCreate.ts
@@ -1,11 +1,17 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
+import { AppError } from "@utils/AppError";
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroup: string) {
     try{
         const storedGroups = await groupsGetAll();
 
+        const groupExists = storedGroups.includes(newGroup);
+
+        if(groupExists){
+            throw new AppError("Já existe grupo cadastrado com esse nome.");
+        }
         const storage = JSON.stringify([...storedGroups, newGroup]);
         await AsyncStorage.setItem(GROUP_COLLECTION, storage);
     }catch(error){
